requirements.txt
fair-esm==2.0.0
matplotlib==3.7.2
numpy==1.6.2
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.10.1
torch==2.0.1
tqdm==4.65.0